Factor out `fine_tune`
