+++ async-openai/src/types/mod.rs
@@ -6,6 +6,7 @@ mod assistant_impls;
 mod audio;
 mod common;
 mod file;
+mod fine_tune;
 mod image;
 mod message;
 mod message_file;
@@ -19,6 +20,7 @@ pub use assistant_file::*;
 pub use audio::*;
 pub use common::*;
 pub use file::*;
+pub use fine_tune::*;
 pub use image::*;
 pub use message::*;
 pub use message_file::*;
